requirements.txt
Django==3.1.4

django-distill==2.4

django-markdownx==3.0.1